+++ thoth-api/src/schema.rs
@@ -553,7 +553,7 @@ table! {
         first_page -> Nullable<Text>,
         last_page -> Nullable<Text>,
         page_interval -> Nullable<Text>,
-        relation_updated_at -> Timestamptz,
+        updated_at_with_relations -> Timestamptz,
     }
 }